producer/requirements.txt
pandas
kafka-python
PyYAML